+++ modules/functions.psm1
@@ -12,6 +12,10 @@ The script will issue an error when something else goes wrong with the file
 .PARAMETER inputFile
 Location of the file containing the valid KQL to execute
 
+.PARAMETER parameters
+Parameters to use for the query. Should be provided as a string containing parameter=value pairs separated by ;.
+In the kql file these parameters should be put with the following syntax %PARAMETER%
+
 .OUTPUTS
 Outputs the result of the KQL query as a PSObject
 
@@ -26,7 +30,10 @@ function ExecuteQuery {
     param
     (
         [Parameter(Mandatory = $true, Position = 0)]
-        [string] $inputFile
+        [string] $inputFile,
+        [Parameter(Mandatory = $false, Position = 1)]
+        [string] $parameters
+
     )
 
     try {
@@ -40,6 +47,13 @@ function ExecuteQuery {
     catch {
         Write-Warning "An uncatched error has occurred. Please investigate"
         Write-Error $error[0]
+    }
+
+    # Replace parameters with their value in the query
+    $parameterValues = ConvertFrom-StringData -StringData ($parameters.Replace(";", "`n"))
+
+    $parameterValues.GetEnumerator() | ForEach-Object {
+        $queryContent = $queryContent.Replace("%$($_.Key.ToUpper())%", $_.Value)
     }
 
     try {
@@ -231,4 +245,443 @@ function ConvertCSVtoMD {
 
     $mdData | Out-File -Append $outPutFile
 
+}
+
+<#
+.SYNOPSIS
+Function that triggers the correct MS Graph function
+
+.DESCRIPTION
+Triggers the correct MS Graph function for the component that is provided as the first parameter.
+The following permission is sufficient to run all current Graph functions:
+- Directory.Read.All
+ 
+.PARAMETER component
+Component to execute the MS Graph function for
+
+.PARAMETER options
+Custom options for the MS Graph function
+
+.OUTPUTS
+Outputs the result of the MS Graph function as a PSObject
+
+.EXAMPLE
+PS> $result = ExecuteMsGraphScript -component "enterpriseApps"
+
+.NOTES
+If no MS Graph function exists for a specific resource type this function returns a warning
+#>
+function ExecuteMsGraphFunction {
+    param (
+        [Parameter(Mandatory = $true, Position = 0)]
+        [string] $component,
+        [Parameter(Mandatory = $false, Position = 1)]
+        [string] $options = ""
+    )
+
+    $result = $null
+
+    switch ($component) {
+        "enterpriseApps" {
+            $result = graphEnterpriseAppsFunction -options "$options"
+        }
+        "managedIdentities" {
+            $result = graphManagedIdentitiesFunction
+        }
+        "directoryRoles" {
+            $result = graphDirectoryRolesFunction
+        }
+        "generalInfo" {
+            $result = graphGeneralInfoFunction
+        }
+        Default {
+            Write-Warning "No MS Graph function defined for component $component"
+        }
+    }
+
+    $result
+}
+
+<#
+.SYNOPSIS
+Function to gather information on existing enterprise apps.
+
+.DESCRIPTION
+Following Microsoft Graph scopes are required for running this script:
+- Application.Read.All
+
+This function will gather the following information on the enterprise apps that exist in the Entra ID tenant
+- Display Name
+- AppId
+- Description
+- Secrets and their expiry date
+- Expired and soon expiring secrets
+- Federated Credentials
+- App Roles
+- Creation Date
+- Owners
+- Sign-in Audience
+
+.PARAMETER options
+Custom options for the enterprise apps function. Should be provided as a string containing key=value pairs separated by ;.
+Possible key values:
+- secretExpiryWarningDays: number of days before end date of secret that it is put in expiring state. Default is 30 days.
+
+.OUTPUTS
+Outputs the results as a PSObject
+
+.EXAMPLE
+PS> Connect-MgGraph -scopes "Application.Read.All"
+PS> $result = graphEnterpriseAppsFunction -options "secretExpiryWarningDays=30"
+
+.NOTES
+The session needs to be connected to Microsoft Graph before running this script.
+#>
+function graphEnterpriseAppsFunction {
+    [Parameter(Mandatory = $false, Position = 0)]
+    [string] $options = ""
+
+    $optionValues = ConvertFrom-StringData -StringData ($options.Replace(";", "`n"))
+
+
+    if ($optionValues.ContainsKey("secretExpiryWarningDays")) {
+        $secretExpiryWarningDays = $optionValues.secretExpiryWarningDays
+        Write-Host "Custom value for secretExpiryWarningDays: $secretExpiryWarningDays"
+    }
+    else {
+        $secretExpiryWarningDays = 30
+    }
+    
+    $context = Get-MgContext
+    $currentDate = Get-Date
+    $expiryWarningDate = (Get-Date).AddDays($secretExpiryWarningDays)
+
+    if ($context) {
+        Write-Host "Gathering info on enterprise apps."
+        $result = [System.Collections.ArrayList]::new()
+        $requiredProperties = "AppId, DisplayName, Description, PasswordCredentials, FederatedIdentityCredentials, AppRoles, CreatedDateTime, Owners, SignInAudience"
+        $applicationList = Get-MgApplication -ExpandProperty FederatedIdentityCredentials -Property $requiredProperties
+        foreach ($application in $applicationList) {
+            # Create calculated members for application
+            ## PassWordCredentials
+            $tempPasswordString = ""
+            $expiryStatus = "None"
+            foreach ($password in $application.PasswordCredentials) {
+                if ($password.DisplayName) {
+                    $tempPasswordString += $password.DisplayName
+                }
+                else {
+                    $tempPasswordString += "<no DisplayName>"
+                }
+                $tempPasswordString += " ($($password.EndDateTime.ToString("dd MMM yyyy hh:mm")))"
+                # Calculate correct status in case of multiple secrets
+                if ($password.EndDateTime -gt $expiryWarningDate) {
+                    $expiryStatus = "OK"
+                }
+                elseif ($password.EndDateTime -gt $currentDate) {
+                    $tempPasswordString += " (EXPIRING)"
+                    if ($expiryStatus -ne "OK") { 
+                        $expiryStatus = "Expiring"
+                    }
+                }
+                else {
+                    $tempPasswordString += " (EXPIRED)"
+                    if ($expiryStatus -eq "None") {
+                        $expiryStatus = "Expired"
+                    }
+                }
+                $tempPasswordString += " | "
+            }
+            if ($tempPasswordString -ne "") {
+                $tempPasswordString = $tempPasswordString -replace ".{3}$"
+            }
+
+            ## FederatedCredentials
+            $tempFederatedCredentials = ""
+            foreach ($credential in $application.FederatedIdentityCredentials) {
+                $tempFederatedCredentials += "Name: $($credential.name), "
+                $tempFederatedCredentials += "Issuer: $($credential.Issuer), "
+                $tempFederatedCredentials += "Subject: $($credential.Subject) | "
+            }
+            if ($tempFederatedCredentials -ne "") {
+                $tempFederatedCredentials = $tempFederatedCredentials -replace ".{3}$"
+            }
+
+            ## AppRoles
+            $tempAppRoles = ""
+            foreach ($role in $application.AppRoles) {
+                $tempAppRoles += "$($role.DisplayName) | "
+            }
+            if ($tempAppRoles -ne "") {
+                $tempAppRoles = $tempAppRoles -replace ".{3}$"
+            }
+
+            ## RBAC roles
+            $servicePrincipalId = (Get-MgServicePrincipal -Filter "AppId eq `'$($application.AppId)`'").Id
+            $queryResults = ExecuteQuery -inputFile "queries/entra/RBACforEntraId.kql" -parameters "ID=$servicePrincipalId"
+            $roles = ""
+            foreach ($role in $queryResults) {
+                $roles += "role: $($role.roleName), scope: $($role.scopeName), type: $($role.scopeType) | "
+            }
+            if ($roles -ne "") {
+                $roles = $roles -replace ".{3}$"
+            }
+  
+
+
+            # Create application custom object to add to array
+            $tempObject = [PSCustomObject]@{
+                DisplayName          = $application.DisplayName
+                AppId                = $application.AppId
+                Description          = $application.Description
+                PasswordCredentials  = $tempPasswordString
+                ExpiryStatus         = $expiryStatus
+                FederatedCredentials = $tempFederatedCredentials
+                AppRoles             = $tempAppRoles
+                CreatedDateTime      = ($application.CreatedDateTime).ToString("dd MMM yyyy hh:mm")
+                Owners               = $application.Owners
+                SignInAudience       = $application.SignInAudience
+                Roles                = $roles
+            }
+            [void]$result.Add($tempObject)
+        }
+    }
+    else {
+        Write-Warning "No connection to Microsoft Graph. Cannot execute the enterpriseApps inventory."
+        $result = $null
+    }
+    
+    return $result
+}
+
+<#
+.SYNOPSIS
+Function to gather information on existing enterprise apps.
+
+.DESCRIPTION
+Following Microsoft Graph scopes are required for running this script:
+- Application.Read.All
+
+This function will gather the following information on the managed identities that exist in the Entra ID tenant
+- Display Name
+- AppId
+- Description
+- Creation Date
+- Owners
+- Resource
+
+.OUTPUTS
+Outputs the results as a PSObject
+
+.EXAMPLE
+PS> Connect-MgGraph -scopes "Application.Read.All"
+PS> $result = graphManagedIdentitiesFunction
+
+.NOTES
+The session needs to be connected to Microsoft Graph before running this script.
+#>
+function graphManagedIdentitiesFunction {
+    
+    $context = Get-MgContext
+
+    if ($context) {
+        Write-Host "Gathering info on managed identities."
+        $result = [System.Collections.ArrayList]::new()
+        $requiredProperties = "Id, DisplayName, AppId, Description, Owners, AlternativeNames, createdDateTime"
+        $managedIdentityList = Get-MgServicePrincipal -filter "ServicePrincipalType eq 'ManagedIdentity'" -Property $requiredProperties
+        foreach ($identity in $managedIdentityList) {
+            # Create calculated members for managed identity
+            ## RBAC roles
+            $queryResults = ExecuteQuery -inputFile "queries/entra/RBACforEntraId.kql" -parameters "ID=$($identity.Id)"
+            $roles = ""
+            foreach ($role in $queryResults) {
+                $roles += "role: $($role.roleName), scope: $($role.scopeName), type: $($role.scopeType) | "
+            }
+            if ($roles -ne "") {
+                $roles = $roles -replace ".{3}$"
+            }
+
+            # Created Date
+            [datetime]$createdDate = $identity.AdditionalProperties.createdDateTime
+            
+            # Create application custom object to add to array
+            $tempObject = [PSCustomObject]@{
+                DisplayName     = $identity.DisplayName
+                AppId           = $identity.AppId
+                Description     = $identity.Description
+                CreatedDateTime = $createdDate.ToString("dd MMM yyyy hh:mm")
+                Owners          = $identity.Owners
+                Resource        = $identity.AlternativeNames[1]
+                Roles           = $roles
+            }
+            [void]$result.Add($tempObject)
+        }
+    }
+    else {
+        Write-Warning "No connection to Microsoft Graph. Cannot execute the managedIdentity inventory."
+        $result = $null
+    }
+    
+    $result
+}
+
+<#
+.SYNOPSIS
+Function to gather information on Entra ID roles and their members
+
+.DESCRIPTION
+Following Microsoft Graph scopes are required for running this script:
+- RoleManagement.Read.Directory
+
+This function will gather the following information on all the Entra ID roles
+- Display Name
+- Members
+
+.OUTPUTS
+Outputs the results as a PSObject
+
+.EXAMPLE
+PS> Connect-MgGraph -scopes "RoleManagement.Read.Directory"
+PS> $result = graphDirectoryRolesFunction
+
+.NOTES
+The session needs to be connected to Microsoft Graph before running this script.
+#>
+function graphDirectoryRolesFunction {
+    
+    $context = Get-MgContext
+
+    if ($context) {
+        Write-Host "Gathering info on Directory Roles"
+        $result = [System.Collections.ArrayList]::new()
+
+        $directoryRoles = Get-MgDirectoryRole -ExpandProperty Members -Property DisplayName,Members
+
+        $cachedUsers = @{}
+        foreach ($role in $directoryRoles) {
+            # Lookup Display names of members
+            $members = ""
+            $amount = 0
+            foreach ($memberId in $role.Members.Id) {
+                if ($cachedUsers.ContainsKey($memberId)) {
+                    $members += "$($cachedUsers[$memberId]), "
+                    $amount += 1
+                }
+                else {
+                    try {
+                        $memberName = (Get-MgDirectoryObject -DirectoryObjectId $memberId -ErrorAction Stop | Select-Object -ExpandProperty AdditionalProperties).displayName
+                        $members += "$memberName, "
+                        $amount += 1
+                    }
+                    catch {
+                        Write-Warning "Something went wrong when collecting the members of group $displayName"
+                    }
+                    $cachedUsers.Add($memberId, $memberName)
+                }
+            }
+            if ($members -ne "") {
+                $members = $members -replace ".{2}$"
+            }
+
+            
+            # Create application custom object to add to array
+            $tempObject = [PSCustomObject]@{
+                DisplayName     = $role.DisplayName
+                NumberOfMembers = $amount
+                Members         = $members
+            }
+            [void]$result.Add($tempObject)
+        }
+    }
+    else {
+        Write-Warning "No connection to Microsoft Graph. Cannot execute the directoryRoles inventory."
+        $result = $null
+    }
+    
+    $result
+}
+
+<#
+.SYNOPSIS
+Function to gather general information regarding the Entra Tenant
+
+.DESCRIPTION
+Following Microsoft Graph scopes are required for running this script:
+- Directory.Read.All
+
+This function will gather the following information on Entra Id
+- Tenant name
+- Tenant ID
+- domains
+
+.OUTPUTS
+Outputs the results as a PSObject
+
+.EXAMPLE
+PS> Connect-MgGraph -scopes "Directory.Read.All"
+PS> $result = graphGeneralInfoFunction
+
+.NOTES
+The session needs to be connected to Microsoft Graph before running this script.
+#>
+function graphGeneralInfoFunction {
+    
+    $context = Get-MgContext
+
+    if ($context) {
+        Write-Host "Gathering info on Entra Tenant"
+        $info = [ordered]@{}
+
+        # Organization
+        $organization = Get-MgOrganization -Property DisplayName,Id
+        $info.Add("Name", $organization.DisplayName)
+        $info.Add("Id", $organization.Id)
+
+        # Domains
+        $domains = Get-MgDomain -Property Id,IsDefault,IsVerified
+        $defaultDomain = ""
+        $verifiedDomains = ""
+        $otherDomains = ""
+        foreach ($domain in $domains) {
+            if($domain.IsDefault){
+                $defaultDomain = $domain.Id
+            }elseif($domain.IsVerified){
+                $verifiedDomains += "$($domain.Id), "
+            }else{
+                $otherDomains += "$($domain.Id), "
+            }
+        }
+        if($verifiedDomains -ne ""){
+            $verifiedDomains = $verifiedDomains -replace ".{2}$"
+        }
+        if($otherDomains -ne ""){
+            $otherDomains = $otherDomains -replace ".{2}$"
+        }
+        $info.Add("Default domain", $defaultDomain)
+        $info.Add("Verified domains", $verifiedDomains)
+        $info.Add("Other domains", $otherDomains)
+        
+        #License
+        $sku = ((Get-MgSubscribedSku).ServicePlans | Where-Object { $_.ServicePlanName -Like 'AAD_PREMIUM*' }).ServicePlanName
+        if($sku -contains "AAD_PREMIUM_P2"){
+            $info.Add("License", "Microsoft Entra ID P2")
+        }elseif ($sku -contains "AAD_PREMIUM") {
+            $info.Add("License", "Microsoft Entra ID P1")
+        }else {
+            $info.Add("License", "Unknown")
+        }
+
+        $result = foreach ($key in $info.Keys) {
+           [PSCustomObject]@{
+            Item = $key
+            Value = $info[$key]
+           }
+        }
+    }
+    else {
+        Write-Warning "No connection to Microsoft Graph. Cannot execute the directoryRoles inventory."
+        $result = $null
+    }
+    
+    $result
 }